Connect to a specific machine when importing
We previously had a bug where, due to the fact that we weren't skipping
DNS registration, `fly pg import` would create a machine that tried to
import from itself, obviously failing. Although the DNS registration bug
is [being fixed](https://github.com/superfly/fly-proxy/pull/1192), I
still think that relying on skipping DNS registration is a little silly,
when we can just guarantee we'll connect to the correct machine using
6PN wizardry. `flyctl` needs to specify a machine ID to connect to
which, as of writing, is just any working machine in the app we want to
import into.